Dataset Extraction Program
This file is used to capture (or) collect the gray-scale image of faces recognize by your camera. In this file there are 30 images are taken and it's accuracy is 30-40%. If you want to increase the accuracy, change dateset 30 to 100 or 200 images(Now 100 image is taken by the camera) then the accuracy is get high.